Remove `new` from the Component trait
The `new` function in the Component trait was not called from generic contexts,
so there is no reason for it to be in the trait. This simplifies things by
decoupling the concept of Component from the tokio UnboundedSender.
More concretely, it allows InputBox::new to take no arguments, which is clearer.